requirements.txt
fastapi
tensorflow==2.5.0
pillow
gunicorn
uvicorn
numpy
python-multipart
protobuf==3.20